+++ chia/wallet/wallet_coin_store.py
@@ -123,7 +123,7 @@ class WalletCoinStore:
                 int(record.coinbase),
                 str(record.coin.puzzle_hash.hex()),
                 str(record.coin.parent_coin_info.hex()),
-                bytes(record.coin.amount),
+                bytes(uint64(record.coin.amount)),
                 record.wallet_type,
                 record.wallet_id,
             ),
